Add stringz module to define some commonly used strings.
This module will eventually contain various useful zero-terminated string functions.

For now, using NULL_Z instead of strPtr(NULL_STR) avoids a strict aliasing warning on RHEL 6.  This is likely a compiler issue, but adding these constants seems like a good idea anyway and we are not going to get a fix in a gcc that old.